Set text entry styles
A text entry should look similar to the address bar of firefox.

There is a bit of a workaround for spinbutton elements. The spinbutton
has an entry element within but it doesn't listen to the entry
selector. To overcome this we style the spinbutton's text color and use
the '*' selector to, restore the textcolor in child elements. For some
reason the entry element listen to the color of the spinbutton.