Show caret in text fields
The default color of the caret in entry/text field elements is the same
as the background. Making it difficult to tell if a text field is
selected or in focus.

Some text fields have icons in them (like a looking glass in a search
field) we want some space between the text and the icons to avoid a
cramped layout.