Darken input fields
As the input field in Firefox's sidebar doesn't listen to the text color
rule of the 'entry' element in gtk.css (unknown why) we'll need the
background of the input field so that the light main text color remain
visible.

If we find a way to target Firefox's sidebar input field we might
reverse this commit.