Remove large margin to text fields
We try to not set large margins on elements unless we need to. Large
margins introduce great risk of messy layouts as this often change the
hight of containing elements and siblings will strech to be as tall
as the containing element. It can be difficult to find the reason for
these weird side effects.