Fix elognated toolbar buttons
The buttons in toolbars in applications such as Inkscape and PCManFM
did not quite look the same. They would sometimes have much greater
hight compared to width. We want them to be fairly squared in all
toolbars. We also want there to be some space between toolbar elements
so we cannot remove all margins.