Use \1 and \2 instead of \[ and \] for escaping
When setting colors in the terminal's prompt we need to tell the
terminal to not count these characters when determening the line
length. We do so by adding escape characters \[ and \] around the color
command. However, if the color command is used in command subsitution,
$(), then we need to use \001 and \002 instead. This version of the
escape characters works in normal use as well. So for the sake of
being consistent we use them everywhere.

ref: https://wiki.archlinux.org/title/Bash/Prompt_customization#Terminfo_escape_sequences