Only use single space symbols
Some symbols (specifically the input label symbol) took two character
spaces rather than one. This can "hide" a space in the prompt line that
leaks into the command line (where we write) and mess up the
formatting. Typically we would see this when scrolling the command
history where the previous commands would jump one step to the left
and when editing the cursor would look like it's editing in one place
but actually editing another.

Basically they behaved as "hidden" full width glyphs when they were
expected to be half or single width (as we thought this was a mono
font.

refs:
https://en.wikipedia.org/wiki/Halfwidth_and_fullwidth_forms
https://en.wikipedia.org/wiki/Duospaced_font
https://en.wikipedia.org/wiki/Monospaced_font