Enable GTK styles override for certain elements
GTK doesn't always seem to listen to the chosen theme, specifically
the gnome-terminal. When setting a background color for the terminal
in the css for the theme in <path>/themes/monster/gtk-3.0/gtk.css the
changes wouldn't show. However one can force these styles by adding
them to ~/.config/gtk-3.0/gtk.css so we add some specific rules to this
css.