Build i3 config from partial configs
In an attempt to minimize the config the user has to do after building
Monster we ask the user to provide a base i3 config to which we'll
append some Monster specific configuration.

We try to warn the user from editing the built config file with a few
scary warning messages.