Add i3bar config
We use i3blocks to create a status bar and we avoid having the user
add extra stuff to the i3 config by merging another partial config.

We realize that users may have other mount points they want to show in
the status bar so we allow them to insert them via the build script.