Darken i3bar backgrounds
When items in the status bar is marked as urgent they will get the
background color of the urgent setting in the bar{} block. Additionally
the status bar is less static in terms of text color, meaning some
status items may use different colors for states like warnings,
critical, etc. For the most part these colors are bright and with the
current background color the contrast would sometimes be too low,
resulting in hidden text. The current colors were chosen for the
workspace icons which are more static.