Organize partial i3 configs
Some of the partial configs where mixing different config areas into
one file, we'd prefer it if they were more specialized for it to be
easier to find things.

Some times the order of the files in the partial configs directory is
important when the configs are merged into one. It should be easy to
add new (but not many) configs between existing ones.