Set color for selected text+background
Selected text in some application, especially in Firefox, only seemed to get
the foreground color from the theme. As the forground color is light this
more often than not would result in the selection not showing or only making
the text slightly lighter.

This answer gave the hint to the solution:
https://stackoverflow.com/a/52501679/2671179