Configure lockscreen
The automatic lockscreen was not configured correctly (missing
dependency) so just installing the theme was not enough to have it
working. Also it was not possible to manually lock the screen.

When the screen was locked it did not match the theme so we use
i3lock-color instead of regular i3lock to get more control over the
lockscreen styling.

To make it easy to setup we provide some sample lockscreens.