Update to valid picom config
When starting picom manually the output suggested that some config
options used were deprecated and no longer valid (probably a new version
of picom). Also when picom loads it's config it relies on the
environment and the values of some variables. We always want to use our
config so we better specify it (this also makes it easier to know which
config is supposedly used as it's visible in the start command).